+++ core/evals/configs/others/heterofl.yml
@@ -8,17 +8,17 @@ ps_ip: 10.0.0.1
 # Note that if we collocate ps and worker on same GPU, then we need to decrease this number of available processes on that GPU by 1
 # E.g., master node has 4 available processes, then 1 for the ps, and worker should be set to: worker:3
 worker_ips:
-    - 10.0.0.2:[2] # worker_ip: [(# processes on gpu) for gpu in available_gpus] eg. 10.0.0.2:[4,4,4,4] This node has 4 gpus, each gpu has 4 processes. 
-    - 10.0.0.3:[2] 
-    - 10.0.0.4:[2]
-    - 10.0.0.5:[2]
+    - 10.0.0.1:[3] # worker_ip: [(# processes on gpu) for gpu in available_gpus] eg. 10.0.0.2:[4,4,4,4] This node has 4 gpus, each gpu has 4 processes. 
+    # - 10.0.0.3:[3] 
+    # - 10.0.0.4:[2]
+    # - 10.0.0.5:[2]
 
 exp_path: $HOME/FedScale/core
 
 # Entry function of executor and aggregator under $exp_path
-executor_entry: ./examples/fed_hetero/customized_executor.py
+executor_entry: ./examples/heterofl/customized_executor.py
 
-aggregator_entry: ./examples/fed_hetero/customized_aggregator.py
+aggregator_entry: ./examples/heterofl/customized_aggregator.py
 
 auth:
     ssh_user: ""
@@ -32,16 +32,16 @@ setup_commands:
 # Default parameters are specified in argParser.py, wherein more description of the parameter can be found
 
 job_conf: 
-    - job_name: fedhet                   # Generate logs under this folder: log_path/job_name/time_stamp
+    - job_name: heterofl                   # Generate logs under this folder: log_path/job_name/time_stamp
     - log_path: $HOME/FedScale/core/evals # Path of log files
     - total_worker: 10                      # Number of participants per round, we use K=100 in our paper, large K will be much slower
     - data_set: cifar10                     # Dataset: openImg, google_speech, stackoverflow
     - data_dir: $HOME/FedScale/dataset/data/    # Path of the dataset
-    - model: resnet_fedhet                      # Models: e.g., shufflenet_v2_x2_0, mobilenet_v2, resnet34, albert-base-v2# - gradient_policy: yogi                 # {"fed-yogi", "fed-prox", "fed-avg"}, "fed-avg" by default
+    - model: resnet_heterofl                      # Models: e.g., shufflenet_v2_x2_0, mobilenet_v2, resnet34, albert-base-v2# - gradient_policy: yogi                 # {"fed-yogi", "fed-prox", "fed-avg"}, "fed-avg" by default
     - eval_interval: 5                     # How many rounds to run a testing on the testing set
     - epochs: 400                          # Number of rounds to run this training. We use 1000 in our paper, while it may converge w/ ~400 rounds
     - filter_less: 0                       # Remove clients w/ less than 21 samples
-    - num_loaders: 5
+    - num_loaders: 2
     - local_steps: 5
     - learning_rate: 0.1
     - batch_size: 10
