--- core/examples/fed_hetero/config.json
@@ -1,5 +1,0 @@
-{
-    "num": 5,
-    "shrinkage": [1, 0.5, 0.25, 0.125, 0.0625],
-    "model_split_mode": "dynamic"
-}
