+++ core/examples/heterofl/customized_fllibs.py
@@ -1,5 +1,5 @@
 import sys, os, logging
-from resnet_fedhet import resnet18
+from resnet_heterofl import resnet18
 import torch
 import torch.nn.functional as F
 import numpy as np
@@ -12,7 +12,7 @@ def init_model():
     global tokenizer
     
     logging.info("Initializing the model ...")
-    if args.model == 'resnet_fedhet':
+    if args.model == 'resnet_heterofl':
         model = resnet18()
     
     return model
